+++ CHANGELOG.md
@@ -17,7 +17,7 @@
     - `unix:<unix timestamp in seconds>`
     - `unix-ms:<unix timestamp in milliseconds>`
 
-- core#1984: Added predefined user group `Data Owner`, members of which can
+- core#1984: Added predefined user group `Repository Manager`, members of which can
     - `pepcli ama query`
     - `pepcli user query`
     - `pepcli query data-size`